chore: add examples project with a local function agent example
