+++ examples/agents/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "paths": {
+      "@aigne/core": ["../../packages/core/src"]
+    }
+  }
+}
