+++ packages/runtime/src/cmd/utils/generate-wrapper-code.ts
@@ -47,13 +47,13 @@ export async function generateWrapperCode(project: ProjectDefinition) {
 
   const index = `\
 import { Runnable } from '@aigne/core';
-import { Runtime, ProjectDefinition } from '@aigne/runtime';
+import { AIGNERuntime, ProjectDefinition } from '@aigne/runtime';
 
 const projectDefinition: ProjectDefinition = ${JSON.stringify(project, null, 2)};
 
 ${generateAgentsInterface(project.runnables, "Runnable")}
 
-export default new Runtime<Agents>({
+export default new AIGNERuntime<Agents>({
   projectDefinition,
 });
 `;
@@ -70,13 +70,13 @@ export default function middleware() {
 
   const client = `\
 import { ProjectDefinition } from '@aigne/runtime';
-import { Agent, Runtime } from '@aigne/runtime/client';
+import { Agent, AIGNERuntime } from '@aigne/runtime/client';
 
 const projectDefinition: ProjectDefinition = ${JSON.stringify(sanitizeProjectDefinition(project), null, 2)};
 
 ${generateAgentsInterface(project.runnables, "Agent")}
 
-export default new Runtime<Agents>({
+export default new AIGNERuntime<Agents>({
   projectDefinition,
 });
 `;
