+++ packages/runtime/src/index.ts
@@ -1,3 +1,4 @@
-export * from "./runtime";
+export * from "./provider/blocklet-api-agent";
 export * from "./provider/blocklet-llm-model";
 export * from "./provider/quickjs-runner";
+export * from "./runtime";
